Posting and viewing posts. Index 75%
We should be able to add a post provided we are in the DB, and submit
the signature one post parameter, and the json encoded data in the
other. This is verified against the claimed users public keys, and
allowed to be posted provided it passes and has the necessary
information.