Posting and updating. Proper sorting.
Posting and updating posts is working almost properly. The problem being
if the title of the post changed. This will result in the post existing
in two places, and also make things be improper for the posts stored in
memory.

The index page was sorting posts backwards, and now does things
correctly. All times are also now stored internally as instances of
Ruby's Time object, instead of as Strings.