Working on the HTML
The HTML should be just about done. I'll add extra tags as I need to,
but no more. From here on out, I think it's going to be basically just
getting the static content of the site together (links and whatnot), and
making it pretty with a bunch of CSS.