Styles, helpers, pagination, sort at startup
Styles are gradually improving. We also have some helper methods that go
along with the next point: Pagination. It's working as one would expect
it to.

I've also added a variable storing all the posts pre-sorted. This way we
don't have to re-sort each time we access them we just sort the existing
ones at startup, and add new posts to the front. Things stay sorted and
we cut down on pointless computation.